+++ R/metrics.R
@@ -1,3 +1,28 @@
+#' List metrics
+#'
+#' List all available performance metrics.
+#'
+#' @export
+#'
+#' @examples
+#' (metrics <- list_metrics())
+#' metrics[metrics$Task == "Multiclass classification", ]
+list_metrics <- function() {
+  data.frame(
+    Metric = c("auc", "error", "logloss", "mauc", "mlogloss", "mse", "r2",
+               "rsquared", "rmse"),
+    Description = c("Area under (ROC) curve", "Misclassification error",
+                    "Log loss", "Multiclass area under (ROC) curve",
+                    "Multiclass log loss", "Mean squared error", "R squared",
+                    "R squared", "Root mean squared error"),
+    Task = c("Binary classification", "Binary classification",
+             "Binary classification", "Multiclass classification",
+             "Multiclass classification", "Regression", "Regression",
+             "Regression", "Regression")
+  )
+}
+
+
 # Regression -------------------------------------------------------------------
 
 #' @keywords internal
@@ -24,24 +49,26 @@ perf_ce <- ModelMetrics::ce
 
 #' @keywords internal
 perf_auc <- function(actual, predicted) {
-  if (NCOL(predicted) != 2L) {
-    stop("Expected a 2 column matrix of predicted class probabilities.")
-  }
+  # if (NCOL(predicted) != 2L) {
+  #   stop("Expected a 2 column matrix of predicted class probabilities.")
+  # }
   ModelMetrics::auc(
     actual = actual,
-    predicted = predicted[, 1L, drop = TRUE]
+    # predicted = predicted[, 1L, drop = TRUE]
+    predicted = predicted
   )
 }
 
 
 #' @keywords internal
 perf_logLoss <- function(actual, predicted) {
-  if (NCOL(predicted) != 2L) {
-    stop("Expected a 2 column matrix of predicted class probabilities.")
-  }
+  # if (NCOL(predicted) != 2L) {
+  #   stop("Expected a 2 column matrix of predicted class probabilities.")
+  # }
   ModelMetrics::logLoss(
     actual = actual,
-    predicted = predicted[, 1L, drop = TRUE]
+    # predicted = predicted[, 1L, drop = TRUE]
+    predicted = predicted
   )
 }